Preparing script to create the selfcal. Modularizing.

+++ reduction/continuum_imaging_AEG.py
@@ -50,7 +50,7 @@ if 'almaimf_rootdir' in locals():
 if os.getenv('ALMAIMF_ROOTDIR') is None:
     try:
         import metadata_tools
-        from metadata_tools import determine_imsize, determine_phasecenter, logprint, json_load_byteified
+        from metadata_tools import determine_phasecenter, logprint, json_load_byteified
         os.environ['ALMAIMF_ROOTDIR'] = os.path.split(metadata_tools.__file__)[0]
     except ImportError:
         raise ValueError("metadata_tools not found on path; make sure to "
@@ -64,13 +64,12 @@ else:
 almaimf_rootdir = os.getenv('ALMAIMF_ROOTDIR')
 
 
-
 from getversion import git_date, git_version
-from metadata_tools import determine_imsize, determine_phasecenter, logprint, json_load_byteified
+from metadata_tools import determine_phasecenter, logprint, json_load_byteified
 #from make_custom_mask import make_custom_mask
 
 from imaging_parameters import imaging_parameters
-
+from fAEG import imagingOptimalParameters
 
 
 from tasks import tclean, exportfits, plotms, split
@@ -120,20 +119,10 @@ if 'only_7m' not in locals():
     else:
         only_7m = False
 
-
-# load the list of continuum MSes from a file
-# (this file has one continuum MS full path, e.g. /path/to/file.ms, per line)
-with open('continuum_mses.txt', 'r') as fh:
-    continuum_mses = [x.strip() for x in fh.readlines()]
-
 do_bsens = [False]
-
 if os.getenv('DO_BSENS') is not None and os.getenv('DO_BSENS').lower() != 'false':
     do_bsens.append(True)
     logprint("Using BSENS measurement set")
-    # continuum_mses += [x.replace('_continuum_merged.cal.ms',
-    #                              '_continuum_merged_bsens.cal.ms')
-    #                    for x in continuum_mses]
 
 if exclude_7m:
     arrayname = '12M'
@@ -142,76 +131,17 @@ elif only_7m:
 else:
     arrayname = '7M12M'
 
+# load the list of continuum MSes from a file
+# (this file has one continuum MS full path, e.g. /path/to/file.ms, per line)
+with open('continuum_mses.txt', 'r') as fh:
+    continuum_mses = [x.strip() for x in fh.readlines()]
 
-continuum_files_per_field = {}
-vis_image_parameters = {}
-bands = set()
-for continuum_ms in continuum_mses:
-    # get EB name
-    ebname =  re.sub('.*(uid[^.]+).*','\\1',os.path.split(continuum_ms)[1])
-    band = metadata['ebs'][ebname]['band']
-    bands = bands | {band}
-    vis_image_parameters[continuum_ms] = {}
-    vis_image_parameters[continuum_ms]['ebname'] = ebname
+opt_imaging_pars = imagingOptimalParameters(continuum_mses,metadata)
+vis_image_parameters = opt_imaging_pars['visBased']
+field_image_parameters = opt_imaging_pars['fieldBased']
+continuum_files_per_field = opt_imaging_pars['filesPerField']
 
-    # allow optional cmdline args to skip one or the other band
-    if os.getenv('BAND_TO_IMAGE'):
-        if band not in os.getenv('BAND_TO_IMAGE'):
-            logprint("Skipping band {0} because it is not in {1}"
-                     .format(band, os.getenv('BAND_TO_IMAGE')),
-                     origin='almaimf_cont_imaging')
-            continue
-        logprint("Imaging only band {0}".format(os.getenv('BAND_TO_IMAGE')),
-                 origin='almaimf_cont_imaging')
-    if not band in continuum_files_per_field:
-        continuum_files_per_field[band] = {}
-    fields = metadata['ebs'][ebname]['fields']
-    for field in fields:
-        if field in continuum_files_per_field[band]:
-            continuum_files_per_field[band][field].append(continuum_ms)
-        else:
-            continuum_files_per_field[band][field] = [continuum_ms]
-
-    field = fields[0]
-    coosys,racen,deccen = determine_phasecenter(ms = continuum_ms, field = field)
-    phasecenter = "{0} {1}deg {2}deg".format(coosys, racen, deccen)
-    vis_image_parameters[continuum_ms]['phasecenter'] = phasecenter
-    (dra,ddec,pixscale) = list(determine_imsize(ms=continuum_ms, field=field,
-                                                phasecenter=(racen,deccen),
-                                                exclude_7m=exclude_7m,
-                                                only_7m=only_7m,
-                                                spw='all',
-                                                pixfraction_of_fwhm=1/8. if only_7m else 1/4.))
-    imsize = [dra, ddec]
-    vis_image_parameters[continuum_ms]['imsize'] = imsize
-    vis_image_parameters[continuum_ms]['cellsize'] = ['{0:0.2f}arcsec'.format(pixscale)] * 2
-    vis_image_parameters[continuum_ms]['cellsize_in_arcsec'] = [floor(pixscale*1000 + 0.5)/1000] * 2  # in arcsec
-    msmd.open(continuum_ms)
-    antennae = ",".join([x for x in msmd.antennanames() if 'CM' not in x])
-    vis_image_parameters[continuum_ms]['12M_antennae'] = antennae
-    antennae = ",".join([x for x in msmd.antennanames() if 'CM' in x])
-    vis_image_parameters[continuum_ms]['7M_antennae'] = antennae
-    vis_image_parameters[continuum_ms]['7M12M_antennae'] = ""
-    msmd.close()
-
-
-field_image_parameters = {}
 bands  = continuum_files_per_field.keys()
-for band in bands:
-    fields = continuum_files_per_field[band].keys()
-    field_image_parameters[band] = {}
-    for field in fields:
-        field_image_parameters[band][field] = {}
-        field_image_parameters[band][field]['imsize'] = [0,0]
-        field_image_parameters[band][field]['cellsize_in_arcsec'] = [1e6,1e6]
-        field_image_parameters[band][field]['phasecenter'] = []
-        for cvis in continuum_files_per_field[band][field]:
-            field_image_parameters[band][field]['imsize'] = max(field_image_parameters[band][field]['imsize'],vis_image_parameters[cvis]['imsize'])
-            field_image_parameters[band][field]['cellsize_in_arcsec'] = min(field_image_parameters[band][field]['cellsize_in_arcsec'],
-                                                                    vis_image_parameters[cvis]['cellsize_in_arcsec'])
-            field_image_parameters[band][field]['cellsize'] = ['{0:0.2f}arcsec'.format(field_image_parameters[band][field]['cellsize_in_arcsec'][0])]*2
-            field_image_parameters[band][field]['phasecenter'].append(vis_image_parameters[cvis]['phasecenter'])
-
 
 # The imaging
 for band in bands:
@@ -246,7 +176,7 @@ for band in bands:
 
                 ## Dirty imaging
                 imname = contimagename+"_robust{0}_dirty".format(robust)
-                if (not os.path.exists(imname+".image.tt0")):
+                if (not os.path.exists(imname+".image.tt0") and not dryRun):
                     logprint("Dirty imaging file {0}".format(imname),
                              origin='almaimf_cont_imaging')
                     #antennae = [vis_image_parameters[cvis][arrayname+'_antennae'] for cvis in continuum_visibilities]
@@ -279,8 +209,11 @@ for band in bands:
                                            "git_date: {0}".format(git_date)])
                     ia.close()
                     exportfits(imname+".image.tt0", imname+".image.tt0.fits", overwrite = True)
+                else:
+                	logprint("Skipping dirty image block. Image={image}, visibilities={vis}".format(image=imname,vis=continuum_visibilities))
+                
                 if onlyDirtyImaging:
-                    logprint("Ony dirty images this run.")
+                    logprint("Only dirty images this run.")
                     continue
  
                 if not 'maskname' in impars:
@@ -335,40 +268,42 @@ for band in bands:
                     #       )""".format(continuum_visibilities,field.encode(),imname,phasecenter,cellsize,imsize,antennae)
                     # sys.exit()
                     rmtables(imname+".mask")
-                    if os.path.exists(imname+".mask"):
-                        os.system("rm -rf "+imname+".mask")
-                    tclean(vis = continuum_visibilities,
-                           field = field.encode(),
-                           imagename = imname,
-                           #phasecenter = phasecenter,
-                           outframe = 'LSRK',
-                           veltype = 'radio',
-                           #usemask='user',
-                           interactive = False,
-                           cell = cell,
-                           imsize = imsize,
-                           antenna = antennae,
-                           pbcor = True,
-                           **impars_thisiter
-                          )
+                    if not dryRun:
+	                    if os.path.exists(imname+".mask"):
+	                        os.system("rm -rf "+imname+".mask")
+	                    tclean(vis = continuum_visibilities,
+	                           field = field.encode(),
+	                           imagename = imname,
+	                           #phasecenter = phasecenter,
+	                           outframe = 'LSRK',
+	                           veltype = 'radio',
+	                           #usemask='user',
+	                           interactive = False,
+	                           cell = cell,
+	                           imsize = imsize,
+	                           antenna = antennae,
+	                           pbcor = True,
+	                           **impars_thisiter
+	                          )
 
-                    stats = imstat(imagename=imname+".image.tt0",algorithm='hinged-fences',fence=2)
-                    mad = stats['medabsdevmed'][0]
-                    rms  = 1.4826* mad
-                    imhead(imname+".image.tt0",mode='put',hdkey='RMS',hdvalue=rms)
- 
-                    #os.system("cp -r {0}".image.tt0" {0}.iter{1}".format(imname,iteracion))
-                    ia.open(imname+".image.tt0")
-                    ia.sethistory(origin='almaimf_cont_imaging',
-                                  history=["{0}: {1}".format(key, val) for key, val in
-                                           impars.items()])
-                    ia.sethistory(origin='almaimf_cont_imaging',
-                                  history=["git_version: {0}".format(git_version),
-                                           "git_date: {0}".format(git_date)])
-                    ia.close()
+	                    stats = imstat(imagename=imname+".image.tt0",algorithm='hinged-fences',fence=2)
+	                    mad = stats['medabsdevmed'][0]
+	                    rms  = 1.4826* mad
+	                    imhead(imname+".image.tt0",mode='put',hdkey='RMS',hdvalue=rms)
+	 
+	                    #os.system("cp -r {0}".image.tt0" {0}.iter{1}".format(imname,iteracion))
+	                    ia.open(imname+".image.tt0")
+	                    ia.sethistory(origin='almaimf_cont_imaging',
+	                                  history=["{0}: {1}".format(key, val) for key, val in
+	                                           impars.items()])
+	                    ia.sethistory(origin='almaimf_cont_imaging',
+	                                  history=["git_version: {0}".format(git_version),
+	                                           "git_date: {0}".format(git_date)])
+	                    ia.close()
 
-                    exportfits(imname+".image.tt0", imname+".image.tt0.fits", overwrite = True)
-                    exportfits(imname+".image.tt0.pbcor", imname+".image.tt0.pbcor.fits", overwrite = True)
-                    exportfits(imname+".model.tt0", imname+".model.tt0.fits", overwrite = True)
-                    exportfits(imname+".residual.tt0", imname+".residual.tt0.fits", overwrite = True)
-
+	                    exportfits(imname+".image.tt0", imname+".image.tt0.fits", overwrite = True)
+	                    exportfits(imname+".image.tt0.pbcor", imname+".image.tt0.pbcor.fits", overwrite = True)
+	                    exportfits(imname+".model.tt0", imname+".model.tt0.fits", overwrite = True)
+	                    exportfits(imname+".residual.tt0", imname+".residual.tt0.fits", overwrite = True)
+                    else:
+                    	logprint("Dry run: skipping imaging of {image} from visibilities {vis}.".format(image=imname,vis=continuum_visibilities))
